Create chat.dart
This is the basic Chat UI - no functionality for now.

The way to call this class:

chat(chatName: "Contact Name", isClass: true/false, adminPrivileges: true/false)

chatName : Displayed name of the contact
isClass : True if this chat will be used for a class, False for direct contacts
adminPriveleges : True if this chat is for a administration user, especially for teachers.